Flyttet inntektskalkulator til egen felles mappe
