+++ src/frontend/Komponenter/Behandling/VedtakOgBeregning/Overgangsstønad/InnvilgeVedtak/BeregnetInntektKalkulator.tsx
@@ -1,58 +1,23 @@
 import { CalculatorIcon } from '@navikt/aksel-icons';
-import { Dropdown, Button, HStack, Tooltip } from '@navikt/ds-react';
+import { Dropdown, Button, Tooltip } from '@navikt/ds-react';
 import React, { FC, useEffect, useRef, useState } from 'react';
-import { BodyShortSmall } from '../../../../../Felles/Visningskomponenter/Tekster';
 import styled from 'styled-components';
-import ForwardedTextField from './ForwardedTextField';
-import { EnsligErrorMessage } from '../../../../../Felles/ErrorMessage/EnsligErrorMessage';
+import InntektsKalkulator from '../../../../../Felles/Kalkulator/Inntektskalkulator';
 
 const StyledDropdownMenu = styled(Dropdown.Menu)`
     width: 23rem;
 `;
 
 const TASTATURTAST_K = 'k';
-const TASTATURTAST_ENTER = 'Enter';
 
 const BeregnetInntektKalkulator: FC<{
     leggTilBeregnetInntektTekstIBegrunnelse: (årsinntekt: number) => void;
 }> = ({ leggTilBeregnetInntektTekstIBegrunnelse }) => {
-    const [årsinntekt, settÅrsinntekt] = useState<string>('');
     const textFieldRef = useRef<HTMLInputElement>(null);
     const [erDropdownÅpen, settErDropdownÅpen] = useState<boolean>(false);
-    const [feilmedling, settFeilmedling] = useState<string>('');
-
-    const oppdaterÅrsinntekt = () => {
-        settFeilmedling('');
-        const årsinntektTall = parseFloat(årsinntekt);
-
-        if (isNaN(årsinntektTall)) {
-            settFeilmedling('Årsinntekt må være et tall');
-            return;
-        }
-
-        leggTilBeregnetInntektTekstIBegrunnelse(årsinntektTall);
-        settÅrsinntekt('');
-        settErDropdownÅpen(false);
-    };
-
-    const handleRegnUtOgLeggTilTekst = (event: React.KeyboardEvent<HTMLInputElement>) => {
-        if (event.key === TASTATURTAST_ENTER) {
-            event.preventDefault();
-            oppdaterÅrsinntekt();
-        }
-    };
 
     const handleOnOpenChange = (erÅpen: boolean) => {
         settErDropdownÅpen(erÅpen);
-    };
-
-    const handleTextFieldOnChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-        const value = event.target.value;
-        settÅrsinntekt(value);
-
-        if (!isNaN(parseFloat(value))) {
-            settFeilmedling('');
-        }
     };
 
     useEffect(() => {
@@ -91,31 +56,11 @@ const BeregnetInntektKalkulator: FC<{
                 </Button>
             </Tooltip>
             <StyledDropdownMenu>
-                <BodyShortSmall>Legg inn årsinntekt for å regne ut +/- 10 prosent.</BodyShortSmall>
-
-                <HStack gap="2" justify="space-between">
-                    <ForwardedTextField
-                        ref={textFieldRef}
-                        placeholder="Årsinntekt"
-                        type="number"
-                        inputMode="numeric"
-                        label=""
-                        size="small"
-                        value={årsinntekt}
-                        onChange={handleTextFieldOnChange}
-                        onKeyDown={handleRegnUtOgLeggTilTekst}
-                    />
-
-                    <Button
-                        type="button"
-                        variant="secondary"
-                        size="xsmall"
-                        onClick={oppdaterÅrsinntekt}
-                    >
-                        Beregn
-                    </Button>
-                    <EnsligErrorMessage>{feilmedling}</EnsligErrorMessage>
-                </HStack>
+                <InntektsKalkulator
+                    leggTilBeregnetInntektTekstIBegrunnelse={
+                        leggTilBeregnetInntektTekstIBegrunnelse
+                    }
+                ></InntektsKalkulator>
             </StyledDropdownMenu>
         </Dropdown>
     );
