+++ src/frontend/Komponenter/Verktøy/InntektskalkulatorSide.tsx
@@ -1,7 +1,8 @@
-import React from 'react';
-import BeregnetInntektKalkulator from '../Behandling/VedtakOgBeregning/Overgangsstønad/InnvilgeVedtak/BeregnetInntektKalkulator';
+import React, { useEffect } from 'react';
 import { genererBeregnetInntektsTekst } from '../../App/hooks/useVerdierForBrev';
+import { Button } from '@navikt/ds-react';
 import useFieldState, { FieldState } from '../../App/hooks/felles/useFieldState';
+import InntektsKalkulator from '../../Felles/Kalkulator/Inntektskalkulator';
 import { EnsligTextArea } from '../../Felles/Input/TekstInput/EnsligTextArea';
 
 export const InntektskalkulatorSide: React.FC = () => {
@@ -10,19 +11,31 @@ export const InntektskalkulatorSide: React.FC = () => {
         const beregnetInntektTekst = genererBeregnetInntektsTekst(årsinntekt);
         inntektBegrunnelseState.setValue((prevState) => prevState + beregnetInntektTekst);
     };
+
+    const nullstillKalkulator = () => {
+        inntektBegrunnelseState.setValue(() => '');
+    };
+
+    useEffect(() => {
+        document.title = 'Inntektskalkulator';
+    }, []);
+
     return (
-        <div>
-            <BeregnetInntektKalkulator
+        <>
+            <InntektsKalkulator
                 leggTilBeregnetInntektTekstIBegrunnelse={leggTilBeregnetInntektTekstIBegrunnelse}
             />
+            <Button type="button" variant="tertiary" size="xsmall" onClick={nullstillKalkulator}>
+                Nullstill
+            </Button>
             <EnsligTextArea
                 value={inntektBegrunnelseState.value}
                 onChange={(event) => {
                     inntektBegrunnelseState.onChange(event);
                 }}
-                label="Beregnet inntekt"
+                label=""
                 maxLength={0}
             />
-        </div>
+        </>
     );
 };
